Move the altivec macros in a stand alone files
